LayoutTests/fast/dom/timer-throttling-aggressiveThermalMitigation-expected.txt

Tests that DOM timers are throttled when aggressive thermal mitigation is enabled.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


Enabling aggressive thermal mitigation
internals.setAggressiveThermalMitigationEnabled(true)
* Nesting: 1
PASS internals.isTimerThrottled(timerHandle) is false
* Nesting: 2
PASS internals.isTimerThrottled(timerHandle) is false
* Nesting: 3
PASS internals.isTimerThrottled(timerHandle) is false
* Nesting: 4
PASS internals.isTimerThrottled(timerHandle) is false
* Nesting: 5
PASS internals.isTimerThrottled(timerHandle) is true
* Nesting: 6
PASS internals.isTimerThrottled(timerHandle) is true

Disabling aggressive thermal mitigation
internals.setAggressiveThermalMitigationEnabled(false)
* Nesting: 7
PASS internals.isTimerThrottled(timerHandle) is false
* Nesting: 8
PASS internals.isTimerThrottled(timerHandle) is false
* Nesting: 9
PASS internals.isTimerThrottled(timerHandle) is false
* Nesting: 10
PASS internals.isTimerThrottled(timerHandle) is false
PASS successfullyParsed is true

TEST COMPLETE
